Fix/admin product file empty s3 filename (#1975)
This PR fixes https://github.com/antiwork/gumroad/issues/1974

### Problem

Product card was not rendering properly when the product had associated
product files, but at least one of them was missing a s3_filename
attribute.

### Solution

Make the `s3_filename` nullable in Typescript, and falback to the
external id when it's nullm so the page can render.
Root cause of why the s3_filename is null in the first place takes more
time to investigate, and it's more relevant and urgent to fix the page
rendering in the meantime for admins.

### AI Disclosure
The fixed was implemented manually.
Tests were generated by AI and edited manually.